services/faster_whisper/pyproject.toml
[build-system]
requires = ["setuptools==82.0.1"]
build-backend = "setuptools.build_meta"

[project]
name = "exocort-faster-whisper-service"
version = "0.1.0"
description = "Standalone Faster Whisper API for audio transcription"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.135.1",
  "python-multipart==0.0.22",
  "uvicorn[standard]==0.42.0",
  "faster-whisper==1.2.1",
  "torch==2.10.0",
  "torchaudio==2.10.0",
  "python-dotenv==1.2.2",
]

[project.optional-dependencies]
dev = ["pytest==9.0.2"]

[project.scripts]
faster-whisper-service = "app.main:main"

[tool.setuptools.packages.find]
include = ["app*", "src*"]
